refactor: extract gateway admin statistics
