+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -24,9 +24,10 @@
 //! # Architecture
 //!
 //! The runtime UI is a single inline HTML string (`admin/html.rs`). The
-//! `dcc-mcp-gateway-admin` crate owns the trace domain, compact, link, and
-//! issue-report projections, Vite/npm build boundary, and embedded asset; this
-//! module owns the gateway-specific HTTP adapter and API handlers.
+//! `dcc-mcp-gateway-admin` crate owns the trace domain, compact, link,
+//! issue-report, and statistics projections, Vite/npm build boundary, and
+//! embedded asset; this module owns the gateway-specific data-source/HTTP
+//! adapters and API handlers.
 //!
 //! # Module layout (PIP-687 split)
 //!
